Add functions to IterativeMap to perform calculations for a particular amount of time, rather than some number of iterations. Since checking the current time is relatively slow, these use the total elapsed time for each call to update an estimated iteration speed it uses next time. This is the same algorithm the animation UI uses.
git-svn-id: http://svn.navi.cx/misc/trunk/fyre@5815 9a66fdd2-3ab9-0310-9df6-91dd0304fdbc

+++ src/explorer.h
@@ -58,7 +58,7 @@ struct _Explorer {
 
   guint idler;
   GTimeVal last_gui_update;
-  gulong render_time;       /* In microseconds */
+  gdouble render_time;       /* In seconds */
 
   gchar* current_tool;
   gboolean tool_active;
